Spawn::method can now be given a proc object that receive as argument a proc object which contains the spawn block. By requiring "spawn/cucumber" you can now
This is added to keep in control of processes that are spawned. Testings what happens during the execution of the block was previously not possible. This addition lets you call the last spawn block whenever you want. E.g. in Cucumber you would use the following step:

When /^The background process is done$/ do
  call_last_spawn_proc
end